# ideahunter/api-spec.yaml
openapi: 3.0.3
info:
  title: IdeaHunter API
  description: |
    AI 기반 비즈니스 아이디어 자동 수집/분석/큐레이션 + Claude Code 구현 프롬프트 생성 API.
    6개 소스(HackerNews, Reddit, ProductHunt, GitHub, PlayStore, AppStore)에서 아이디어를 수집하고,
    Groq AI로 분석한 뒤 Claude Code 프롬프트를 자동 생성합니다.
  version: 1.0.0
  contact:
    name: IdeaHunter
servers:
  - url: https://ideahunter.vercel.app
    description: 프로덕션 서버
  - url: http://localhost:3000
    description: 로컬 개발 서버
paths:
...
  /api/ideas:
    get:
      summary: 아이디어 목록 조회
      description: |
        AI 분석이 완료된(summary_ko가 존재하는) 아이디어 목록을 반환합니다.
        정렬, 소스 필터, 태그 필터, 페이지네이션을 지원합니다.
        - trend_score 정렬: 트렌드 점수 내림차순 (기본값)
        - latest 정렬: 수집일시 내림차순
        - ait_score 정렬: platform_analysis가 있는 아이디어만 필터링 후 ait_score 내림차순
      operationId: getIdeas
      tags: [아이디어]
      parameters:
        - name: sort
          in: query
          description: |
            정렬 기준.
            - trend_score: 트렌드 점수 내림차순 (기본값)
            - latest: 수집일 내림차순
            - ait_score: 앱인토스 적합 점수 내림차순 (platform_analysis 존재하는 항목만)
          schema:
            type: string
            enum: [trend_score, latest, ait_score]
            default: trend_score
        - name: source
          in: query
          description: 소스 필터. 허용된 값만 사용 가능.
          schema:
            type: string
            enum: [hackernews, reddit, producthunt, github, playstore, appstore]
        - name: tag
          in: query
          description: 태그 필터. 영문/숫자/공백/하이픈만 허용, 최대 50자.
          schema:
            type: string
            pattern: '^[\w\s-]{1,50}$'
        - name: page
          in: query
          description: 페이지 번호 (0부터 시작, 최대 100)
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 0
        - name: limit
          in: query
          description: 페이지당 항목 수 (1~50)
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: 아이디어 목록 반환 성공
          headers:
            Cache-Control:
              description: CDN 캐시 설정 (5분)
              schema:
                type: string
                example: 's-maxage=300, stale-while-revalidate'
          content:
            application/json:
              schema:
                type: object
                required: [ideas]
                properties:
                  ideas:
                    type: array
                    items:
                      $ref: '#/components/schemas/Idea'
        '400':
          description: 잘못된 요청 파라미터 (source 또는 tag 형식 오류)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 오류 (DB 쿼리 실패)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ──────────────────────────────────────────────
  # Claude Code 프롬프트 생성
  # ──────────────────────────────────────────────
...
  /api/cron/collect:
    get:
      summary: 아이디어 자동 수집 (크론 작업)
      description: |
        7개 소스(HackerNews, Reddit, ProductHunt, GitHub, PlayStore, AppStore, AppBrain)에서
        트렌딩 아이디어를 수집합니다. 각 소스별 최대 10개씩 수집하며, 중복 URL은 건너뜁니다.

        **수집 프로세스**:
        1. 소스별 수집기 실행 → 최대 10개 항목 추출
        2. source_url 기준 중복 체크 (배치)
        3. Groq AI 분석 (summary_ko, difficulty, revenue_potential 등)
        4. trend_score 계산
        5. Supabase ideas 테이블 배치 삽입
        6. collect_logs에 소스별 + 전체 수집 결과 기록

        CRON_SECRET Bearer 인증이 필요합니다.
      operationId: cronCollect
      tags: [크론]
      security:
        - CronAuth: []
      responses:
        '200':
          description: 수집 완료
          content:
            application/json:
              schema:
                type: object
                required: [collected]
                properties:
                  collected:
                    type: integer
                    description: 총 수집 및 삽입된 아이디어 수
                    example: 42
        '401':
          description: 인증 실패 (CRON_SECRET 불일치)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ──────────────────────────────────────────────
  # 크론: 일일 다이제스트 생성
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    CronAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        크론 작업 인증. "Bearer {CRON_SECRET}" 형식으로 전달합니다.
        환경변수 CRON_SECRET과 일치해야 합니다.
  schemas:
...
    Idea:
      type: object
      description: AI 분석이 포함된 비즈니스 아이디어
      required: [id, title, source, score, comment_count, collected_at, trend_score, implementation_status]
      properties:
        id:
          type: string
          format: uuid
          description: 아이디어 고유 식별자
        title:
          type: string
          description: 아이디어 제목
        description:
          type: string
          nullable: true
          description: 원본 설명 텍스트
        source:
          type: string
          enum: [hackernews, reddit, producthunt, github, playstore, appstore]
          description: 수집 소스 플랫폼
        source_url:
          type: string
          format: uri
          description: 원본 게시물 URL
        score:
          type: integer
          description: 소스 플랫폼의 점수/업보트 수
          default: 0
        comment_count:
          type: integer
          description: 댓글/토론 수
          default: 0
        collected_at:
          type: string
          format: date-time
          description: 수집 시각 (UTC)
        summary_ko:
          type: string
          nullable: true
          description: AI 생성 한국어 요약 (Groq llama-3.1-8b-instant)
        market_size:
          type: string
          nullable: true
          description: 시장 규모 추정 텍스트
          example: '국내 약 500억원 규모, 글로벌 $2B'
        difficulty:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
          description: 구현 난이도 (1=매우 쉬움, 5=매우 어려움)
        revenue_potential:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
          description: 수익 잠재력 (1=낮음, 5=매우 높음)
        competition:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
          description: 경쟁 강도 (1=블루오션, 5=레드오션)
        recommended_stack:
          type: array
          nullable: true
          items:
            type: string
          description: 추천 기술 스택
          example: ['Next.js', 'Supabase', 'Tailwind CSS']
        mvp_days:
          type: integer
          nullable: true
          description: MVP 예상 개발 일수
          example: 14
        tags:
          type: array
          nullable: true
          items:
            type: string
          description: 카테고리 태그
          example: ['SaaS', 'AI', 'B2B']
        trend_score:
          type: number
          format: float
          description: 종합 트렌드 점수 (score, comment_count, AI 분석 결과 기반)
          default: 0
        generated_prompts:
          nullable: true
          description: |
            생성된 Claude Code 프롬프트 데이터.
            구조: { default?: GeneratedPrompts, appintoss?: GeneratedPrompts }
            또는 레거시 호환 flat GeneratedPrompts.
          oneOf:
            - $ref: '#/components/schemas/PromptStore'
            - $ref: '#/components/schemas/GeneratedPrompts'
        platform_analysis:
          nullable: true
          description: 앱인토스 플랫폼 적합도 분석 결과 (JSONB)
          allOf:
            - $ref: '#/components/schemas/PlatformAnalysis'
        implementation_status:
          type: string
          enum: [pending, generating, analyzing, done, error]
          default: pending
          description: |
            프롬프트/분석 상태.
            - pending: 대기 중
            - generating: 프롬프트 생성 중 (잠금)
            - analyzing: AIT 분석 중 (잠금)
            - done: 완료
            - error: 실패
        raw_data:
          type: object
          nullable: true
          additionalProperties: true
          description: 수집 원본 데이터 (소스별 상이)

    # ──────────────────────────────────────────
    # 프롬프트 저장 구조
    # ──────────────────────────────────────────
    PromptStore:
      type: object
      description: |
        플랫폼별 프롬프트 저장 구조.
        generated_prompts JSONB 컬럼에 저장되며 default/appintoss 키로 구분합니다.
      properties:
        default:
          $ref: '#/components/schemas/GeneratedPrompts'
        appintoss:
          $ref: '#/components/schemas/GeneratedPrompts'

    # ──────────────────────────────────────────
    # 생성된 프롬프트
    # ──────────────────────────────────────────
    GeneratedPrompts:
      type: object
      description: Claude Code에 입력할 구현 프롬프트 세트
      required: [project_name, overview, master_prompt, phases, tech_stack, free_services]
      properties:
        project_name:
          type: string
          description: 프로젝트 이름 (영문 케밥 케이스)
          example: smart-finance-tracker
        overview:
          type: string
          description: 프로젝트 개요 (한국어)
        master_prompt:
          type: string
          description: Claude Code에 직접 입력할 마스터 프롬프트 (전체 구현 지시)
        phases:
          type: array
          description: 단계별 구현 프롬프트 (보통 4-6단계)
          items:
            type: object
            required: [step, title, prompt]
            properties:
              step:
                type: integer
                description: 단계 번호 (1부터 시작)
              title:
                type: string
                description: 단계 제목
                example: '프로젝트 초기화 + DB 스키마'
              prompt:
                type: string
                description: 해당 단계의 Claude Code 프롬프트
        tech_stack:
          type: array
          items:
            type: string
          description: 사용할 기술 스택 목록
          example: ['Next.js', 'Supabase', 'Tailwind CSS', 'Vercel']
        free_services:
          type: array
          items:
            type: string
          description: 활용 가능한 무료 서비스/티어 목록
          example: ['Supabase Free', 'Vercel Hobby', 'Groq Free']

    # ──────────────────────────────────────────
    # 앱인토스 전용 프롬프트 (확장)
    # ──────────────────────────────────────────
...
    PlatformAnalysis:
      type: object
      description: 앱인토스(토스 미니앱) 플랫폼 적합도 분석 결과
      required: [ait_score, ait_category, ait_ad_revenue_estimate, ait_target_type]
      properties:
        ait_score:
          type: number
          format: float
          description: 앱인토스 적합도 점수 (0-100). 70점 이상이면 적합.
          example: 85
        ait_category:
          type: string
          description: 앱인토스 카테고리 분류
          example: '생활/편의'
        ait_blocked_reason:
          type: string
          nullable: true
          description: 부적합 사유. 적합한 경우 null.
          example: null
        ait_ad_revenue_estimate:
          type: string
          description: 월 광고 수익 추정
          example: '월 50-100만원'
        ait_target_type:
          type: string
          enum: [webview, react-native]
          description: 권장 구현 방식

    # ──────────────────────────────────────────
    # 다이제스트
    # ──────────────────────────────────────────
...
    ErrorResponse:
      type: object
      description: API 에러 응답 공통 형식
      required: [error]
      properties:
        error:
          type: string
          description: 에러 메시지
          example: 'Not found'